Fix mint

+++ math/ModInt.hpp
@@ -21,8 +21,8 @@ template<unsigned int mod> class StaticModInt : StaticModIntBase {
   public:
     StaticModInt() : StaticModInt(0) {}
     template<class T, typename std::enable_if<std::is_integral<T>::value>::type* = nullptr> StaticModInt(T v) {
-        v %= mod;
-        if (v < 0) v += mod;
+        v %= (long long)mod;
+        if (v < 0) v += (long long)mod;
         val = v;
     }
     unsigned int get() const { return val; }
